# openapi.yaml
openapi: "3.1.0"
info:
  title: TFT Agent Set 17 API
  description: |
    TFT Set 17 "Space Gods" AI agent application.
    Provides SSE streaming chat, JWT authentication, conversation management,
    and internal tool dispatch / RAG / graph query endpoints.
  version: "0.1.0"
  contact:
    name: TFT Agent Team
  license:
    name: MIT

servers:
  - url: http://localhost:8000
    description: Local development server

security:
  - BearerAuth: []

paths:
  /v1/chat/completions:
    post:
      operationId: chatCompletions
      summary: Streaming chat completion (SSE)
      description: |
        Accepts a user question and optional direction hint.
        Returns a Server-Sent Events stream with incremental agent response stages:
        understanding, tool_selection, tool_execution, composing.
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: SSE streaming response
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
          headers:
            Cache-Control:
              schema:
                type: string
                example: "no-cache"
            Connection:
              schema:
                type: string
                example: "keep-alive"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

  /v1/health:
    get:
      operationId: healthCheck
      summary: Health check
      tags:
        - System
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                required:
                  - status

  /v1/metrics:
    get:
      operationId: metrics
      summary: Prometheus metrics endpoint
      description: Returns metrics in Prometheus text exposition format.
      tags:
        - System
      security: []
      responses:
        "200":
          description: Prometheus text format metrics
          content:
            text/plain:
              schema:
                type: string

  /auth/register:
    post:
      operationId: register
      summary: Register a new user
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  phone:
                    type: string
                required:
                  - user_id
                  - phone
        "409":
          description: Phone number already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  /auth/login:
    post:
      operationId: login
      summary: Authenticate and obtain tokens
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /conversations:
    get:
      operationId: listConversations
      summary: List user conversations
      tags:
        - Conversations
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Conversation"
                  total:
                    type: integer
                required:
                  - items
                  - total
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createConversation
      summary: Create a new conversation
      tags:
        - Conversations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /conversations/{conversation_id}/messages:
    get:
      operationId: listMessages
      summary: List messages in a conversation
      tags:
        - Conversations
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
                  total:
                    type: integer
                required:
                  - items
                  - total
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Conversation not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login

  schemas:
    Direction:
      type: string
      enum:
        - "\u63a8\u8350\u9635\u5bb9"
        - "\u63a8\u8350\u88c5\u5907"
        - "\u67e5\u4e13\u5c5e"
        - "\u68c0\u7d22\u88c5\u5907"
      description: |
        Hint for the agent about the user's intent:
        - \u63a8\u8350\u9635\u5bb9: Recommend team composition
        - \u63a8\u8350\u88c5\u5907: Recommend items/equipment
        - \u67e5\u4e13\u5c5e: Look up exclusive/champion-specific info
        - \u68c0\u7d22\u88c5\u5907: Search/query items/equipment

    AskRequest:
      type: object
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 2000
          description: User's question in natural language
        direction:
          $ref: "#/components/schemas/Direction"
        conversation_id:
          type: string
          format: uuid
          description: Optional conversation ID to continue a session
      required:
        - question

    AskResponse:
      type: object
      properties:
        answer:
          type: string
          description: The agent's composed answer
        direction:
          $ref: "#/components/schemas/Direction"
        tools_used:
          type: array
          items:
            type: string
          description: List of tool names invoked during processing
        conversation_id:
          type: string
          format: uuid
        latency_ms:
          type: number
          description: Total processing latency in milliseconds
      required:
        - answer

    SSEEvent:
      type: object
      description: |
        A Server-Sent Event. The `event` field indicates the stage:
        - understanding: NLU / intent classification
        - tool_selection: deciding which tools to call
        - tool_execution: tool dispatch results
        - composing: LLM composing the final answer
        - done: final aggregated response
      properties:
        event:
          type: string
          enum:
            - understanding
            - tool_selection
            - tool_execution
            - composing
            - done
        data:
          type: string
          description: JSON-encoded payload for this stage
      required:
        - event
        - data

    RegisterRequest:
      type: object
      properties:
        phone:
          type: string
          pattern: "^1[3-9]\\d{9}$"
          description: Chinese mobile phone number
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
      required:
        - phone
        - password

    LoginRequest:
      type: object
      properties:
        phone:
          type: string
          pattern: "^1[3-9]\\d{9}$"
        password:
          type: string
          format: password
      required:
        - phone
        - password

    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: JWT refresh token
      required:
        - refresh_token

    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (short-lived)
        refresh_token:
          type: string
          description: JWT refresh token (long-lived)
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Access token TTL in seconds
          example: 3600
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in

    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer
      required:
        - id
        - title
        - created_at
        - updated_at

    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - user
            - assistant
        content:
          type: string
        created_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
          description: Optional metadata (tools used, latency, etc.)
      required:
        - id
        - conversation_id
        - role
        - content
        - created_at

    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string
      required:
        - detail

  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"